Refactered remaining APIs.  All back to normal now.
